Rename state2 to state (remove old state)
