+++ pkg/config/gist.go
@@ -10,7 +10,7 @@ import (
 	"path/filepath"
 
 	"github.com/b4b4r07/afx/pkg/errors"
-	"github.com/b4b4r07/afx/pkg/state2"
+	"github.com/b4b4r07/afx/pkg/state"
 	git "gopkg.in/src-d/go-git.v4"
 )
 
@@ -198,6 +198,6 @@ func (c Gist) GetDependsOn() []string {
 	return c.DependsOn
 }
 
-func (c Gist) GetResource() state2.Resource {
+func (c Gist) GetResource() state.Resource {
 	return getResource(c)
 }
