+++ pkg/config/github.go
@@ -20,7 +20,7 @@ import (
 	"github.com/b4b4r07/afx/pkg/data"
 	"github.com/b4b4r07/afx/pkg/errors"
 	"github.com/b4b4r07/afx/pkg/logging"
-	"github.com/b4b4r07/afx/pkg/state2"
+	"github.com/b4b4r07/afx/pkg/state"
 	"github.com/b4b4r07/afx/pkg/templates"
 	"github.com/google/go-github/github"
 	"github.com/mholt/archiver"
@@ -682,6 +682,6 @@ func (c GitHub) GetDependsOn() []string {
 	return c.DependsOn
 }
 
-func (c GitHub) GetResource() state2.Resource {
+func (c GitHub) GetResource() state.Resource {
 	return getResource(c)
 }
