+++ pkg/config/http.go
@@ -11,7 +11,7 @@ import (
 	"path/filepath"
 
 	"github.com/b4b4r07/afx/pkg/errors"
-	"github.com/b4b4r07/afx/pkg/state2"
+	"github.com/b4b4r07/afx/pkg/state"
 	"github.com/h2non/filetype"
 	"github.com/mholt/archiver"
 )
@@ -220,6 +220,6 @@ func (c HTTP) GetDependsOn() []string {
 	return c.DependsOn
 }
 
-func (c HTTP) GetResource() state2.Resource {
+func (c HTTP) GetResource() state.Resource {
 	return getResource(c)
 }
